Add tooling for posts existing in own folder.
After a point, having all the posts in one folder becomes a major mess.
This has been remedied by storing the markdown for the posts in a
subfolder (`posts/`), and by updating the Makefile and archive-generator
scripts to account for the change.

Also included in this commit: The Makefile does not leave HTML detritus
in the posts/ folder. It moves the html files to the `html/` output
folder.